rename recurrent to recurrent_transducers
