+++ test/test_beam_search.py
@@ -9,7 +9,7 @@ from xnmt.bridges import CopyBridge
 from xnmt.decoders import AutoRegressiveDecoder
 from xnmt.embedders import SimpleWordEmbedder
 from xnmt.input_readers import PlainTextReader
-from xnmt.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from xnmt.recurrent_transducers import UniLSTMSeqTransducer, BiLSTMSeqTransducer
 from xnmt.loss_calculators import AutoRegressiveMLELoss
 from xnmt.param_collections import ParamManager
 from xnmt.transforms import NonLinear
