+++ test/test_decoders.py
@@ -10,7 +10,7 @@ from xnmt.decoders import AutoRegressiveDecoder
 from xnmt.embedders import SimpleWordEmbedder
 from xnmt.input_readers import PlainTextReader
 from xnmt.loss_calculators import AutoRegressiveMLELoss
-from xnmt.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from xnmt.recurrent_transducers import UniLSTMSeqTransducer, BiLSTMSeqTransducer
 from xnmt.param_collections import ParamManager
 from xnmt.transforms import NonLinear
 from xnmt.translators import DefaultTranslator
