+++ test/test_hyperparams.py
@@ -8,7 +8,7 @@ from xnmt.embedders import SimpleWordEmbedder
 from xnmt.eval_tasks import LossEvalTask
 from xnmt import events
 from xnmt.input_readers import PlainTextReader
-from xnmt.recurrent import UniLSTMSeqTransducer, BiLSTMSeqTransducer
+from xnmt.recurrent_transducers import UniLSTMSeqTransducer, BiLSTMSeqTransducer
 from xnmt.loss_calculators import AutoRegressiveMLELoss
 from xnmt.optimizers import AdamTrainer
 from xnmt.param_collections import ParamManager
