+++ xnmt/__init__.py
@@ -35,7 +35,7 @@ import xnmt.infererences
 import xnmt.input
 import xnmt.input_readers
 import xnmt.language_models
-import xnmt.recurrent
+import xnmt.recurrent_transducers
 import xnmt.model_base
 import xnmt.optimizers
 import xnmt.param_initializers
