+++ xnmt/language_models.py
@@ -1,7 +1,7 @@
 import dynet as dy
 import numpy as np
 
-from xnmt import batchers, embedders, events, input_readers, losses, recurrent, model_base, scorers, transducers, transforms
+from xnmt import batchers, embedders, events, input_readers, losses, recurrent_transducers, model_base, scorers, transducers, transforms
 from xnmt.persistence import serializable_init, Serializable, bare
 
 class LanguageModel(model_base.ConditionedModel, model_base.EventTrigger, Serializable):
@@ -23,7 +23,7 @@ class LanguageModel(model_base.ConditionedModel, model_base.EventTrigger, Serial
   def __init__(self,
                src_reader:input_readers.InputReader,
                src_embedder:embedders.Embedder=bare(embedders.SimpleWordEmbedder),
-               rnn:transducers.SeqTransducer=bare(recurrent.UniLSTMSeqTransducer),
+               rnn:transducers.SeqTransducer=bare(recurrent_transducers.UniLSTMSeqTransducer),
                transform:transforms.Transform=bare(transforms.NonLinear),
                scorer:scorers.Scorer=bare(scorers.Softmax)):
     super().__init__(src_reader=src_reader, trg_reader=src_reader)
