+++ xnmt/pyramidal.py
@@ -1,7 +1,7 @@
 import dynet as dy
 from typing import List
 
-from xnmt.recurrent import UniLSTMSeqTransducer
+from xnmt.recurrent_transducers import UniLSTMSeqTransducer
 from xnmt.expression_seqs import ExpressionSequence, ReversedExpressionSequence
 from xnmt.persistence import serializable_init, Serializable, Ref
 from xnmt.events import register_xnmt_handler, handle_xnmt_event
