+++ xnmt/recurrent_transducers.py
